Added basic "check" command infrastructure test
This commit is a first draft of a "check" commandline argument command
that checks a list of "ICheck" implementing instances. There's a
CheckRunner that collects the IResult instances of the ICheck instances
and generates an IReport instance, that is used to output the messages
from the Checks using a simple ConsoleFormatter. The Results contain
messages and settings, thus every ICheck may add IMessages with a text and
certain severity level to the Result as well as emit some Settings that
can be later returned in a meaningful way by the CheckCommand.