Added rudimentary config file handling
CheckCommand uses the new DefaultConfigHandler and there will be readers
that can read different config file formats. There's a guesser with
support for xml, json and php config files as well if a config directory
is given instead of a regular file.

This is a work-in-progress. The PHAR is outdated and will probably be
removed and only added for releases.