added complex type to xsd for xi:include support.
when xi:including environaut configs that xi:include other environaut
configs, the schema validation fails.
this is due to the fact that the include processor expands all included
entities adding in a xml:base attribute.
the current environaut schema does not allow for the occurence of such
an attribute on the name, description and information nodes of the
environautType.
more on the topic can be found here:
http://norman.walsh.name/2005/04/01/xinclude